Fix stale Combine→applyData wiring and alarm delete in Recent/Nearby/Search VMs
   - NearbyStopsViewController: $stops sink uses .receive(on:) and calls
     listView.applyData() directly instead of via searchFilter.
   - RecentStopsViewController: CombineLatest($alarms,$recentStops) sink uses
     .receive(on:) so items(for:) reads committed values.
   - RecentStopsViewModel: drop deletionError; log remote alarm-delete
     failures via Logger.error instead of surfacing a modal for a
     non-actionable background operation.
   - UserDataStore.delete(alarm:): match on url instead of full equality;
     Alarm.isEqual compares Dates that lose precision on the UserDefaults
     round-trip, so delete was a no-op.
   - Tests: extract shared CountingDataLoader to Helpers/Mocks; drop the
     deletionError-specific tests now that the path logs only.
