+++ OBAKitTests/Search/SearchViewModelTests.swift
@@ -15,27 +15,6 @@ import Nimble
 
 // swiftlint:disable force_try
 
-// MARK: - CountingDataLoader
-
-private class CountingDataLoader: NSObject, URLDataLoader {
-    let inner: MockDataLoader
-    private(set) var callCount = 0
-
-    init(_ inner: MockDataLoader) { self.inner = inner }
-
-    func dataTask(with request: URLRequest, completionHandler: @escaping (Data?, URLResponse?, Error?) -> Void) -> URLSessionDataTask {
-        inner.dataTask(with: request, completionHandler: completionHandler)
-    }
-
-    func data(for request: URLRequest) async throws -> (Data, URLResponse) {
-        callCount += 1
-        await Task.yield()
-        return try await inner.data(for: request)
-    }
-}
-
-// MARK: -
-
 class SearchViewModelTests: OBATestCase {
 
     // MARK: - Helpers
